# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ellipsonic OpsHub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: unread_only
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Paginated notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPage"
  /notifications/unread-count:
    get:
      operationId: getUnreadNotificationCount
      tags: [notifications]
      summary: Get unread notification count
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Unread count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnreadCount"
  /notifications/read-all:
    put:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /notifications/{id}/read:
    put:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  # ───── ACTIVITY LOGS ─────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # ── Auth / User ──
...
    Notification:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        type:
          type: string
          enum: [reminder, approval, rejected, deadline, announcement, needs_changes]
        title:
          type: string
        message:
          type: string
        entity_type:
          type: ["string", "null"]
        entity_id:
          type: ["string", "null"]
        is_read:
          type: boolean
        created_at:
          type: string
      required: [id, user_id, type, title, message, is_read, created_at]
    NotificationPage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        total_pages:
          type: integer
        unread_count:
          type: integer
      required: [data, total, page, limit, total_pages, unread_count]
    UnreadCount:
      type: object
      properties:
        count:
          type: integer
      required: [count]

    # ── Activity Logs ──
...